+++ src/main/java/com/caoccao/javet/interop/converters/IJavetConverter.java
@@ -23,42 +23,117 @@ import com.caoccao.javet.interop.V8Runtime;
 import com.caoccao.javet.utils.JavetResourceUtils;
 import com.caoccao.javet.values.V8Value;
 
+/**
+ * The interface Javet converter.
+ */
 @SuppressWarnings("unchecked")
 public interface IJavetConverter {
+    /**
+     * Gets default boolean.
+     *
+     * @return the default boolean
+     */
     default boolean getDefaultBoolean() {
         return false;
     }
 
+    /**
+     * Gets default byte.
+     *
+     * @return the default byte
+     */
     default byte getDefaultByte() {
         return (byte) 0;
     }
 
+    /**
+     * Gets default char.
+     *
+     * @return the default char
+     */
     default char getDefaultChar() {
         return '\0';
     }
 
+    /**
+     * Gets default double.
+     *
+     * @return the default double
+     */
     default double getDefaultDouble() {
         return 0D;
     }
 
+    /**
+     * Gets default float.
+     *
+     * @return the default float
+     */
     default float getDefaultFloat() {
         return 0F;
     }
 
+    /**
+     * Gets default int.
+     *
+     * @return the default int
+     */
     default int getDefaultInt() {
         return 0;
     }
 
+    /**
+     * Gets default long.
+     *
+     * @return the default long
+     */
     default long getDefaultLong() {
         return 0L;
     }
 
+    /**
+     * Gets default short.
+     *
+     * @return the default short
+     */
     default short getDefaultShort() {
         return (short) 0;
     }
 
+    /**
+     * Gets max depth.
+     *
+     * @return the max depth
+     */
+    int getMaxDepth();
+
+    /**
+     * Sets max depth.
+     *
+     * @param maxDepth the max depth
+     */
+    void setMaxDepth(int maxDepth);
+
+    /**
+     * To object.
+     *
+     * Don't override this function, instead, override the one with depth as argument
+     * for circular structure detection.
+     *
+     * @param v8Value the V8 value
+     * @return the object
+     * @throws JavetException the javet exception
+     */
     Object toObject(V8Value v8Value) throws JavetException;
 
+    /**
+     * To object.
+     *
+     * @param v8Value   the V8 value
+     * @param autoClose the auto close
+     * @return the object
+     * @throws JavetException the javet exception
+     */
     default Object toObject(V8Value v8Value, boolean autoClose) throws JavetException {
         if (autoClose) {
             try {
@@ -71,6 +146,18 @@ public interface IJavetConverter {
         }
     }
 
+    /**
+     * To V8 value.
+     *
+     * Don't override this function, instead, override the one with depth as argument
+     * for circular structure detection.
+     *
+     * @param <T>       the type parameter
+     * @param v8Runtime the V8 runtime
+     * @param object    the object
+     * @return the V8 value
+     * @throws JavetException the javet exception
+     */
     @CheckReturnValue
     <T extends V8Value> T toV8Value(V8Runtime v8Runtime, Object object) throws JavetException;
 }