+++ src/main/java/com/caoccao/javet/interop/converters/JavetObjectConverter.java
@@ -47,25 +47,25 @@ public class JavetObjectConverter extends JavetPrimitiveConverter {
     }
 
     @Override
-    public Object toObject(V8Value v8Value) throws JavetException {
-        Object returnObject = super.toObject(v8Value);
+    protected Object toObject(V8Value v8Value, final int depth) throws JavetException {
+        Object returnObject = super.toObject(v8Value, depth);
         if (!(returnObject instanceof V8Value)) {
             return returnObject;
         }
         if (v8Value instanceof V8ValueArray) {
             V8ValueArray v8ValueArray = (V8ValueArray) v8Value;
             List<Object> list = new ArrayList<>();
-            v8ValueArray.forEach(value -> list.add(toObject(value)));
+            v8ValueArray.forEach(value -> list.add(toObject(value, depth + 1)));
             return list;
         } else if (v8Value instanceof V8ValueSet) {
             V8ValueSet v8ValueSet = (V8ValueSet) v8Value;
             HashSet<Object> set = new HashSet<>();
-            v8ValueSet.forEach(key -> set.add(toObject(key)));
+            v8ValueSet.forEach(key -> set.add(toObject(key, depth + 1)));
             return set;
         } else if (v8Value instanceof V8ValueMap) {
             V8ValueMap v8ValueMap = (V8ValueMap) v8Value;
             Map<String, Object> map = createEntityMap();
-            v8ValueMap.forEach((V8Value key, V8Value value) -> map.put(key.toString(), toObject(value)));
+            v8ValueMap.forEach((V8Value key, V8Value value) -> map.put(key.toString(), toObject(value, depth + 1)));
             return map;
         } else if (v8Value instanceof V8ValueTypedArray) {
             V8ValueTypedArray v8ValueTypedArray = (V8ValueTypedArray) v8Value;
@@ -103,7 +103,7 @@ public class JavetObjectConverter extends JavetPrimitiveConverter {
                     }
                 } else {
                     try (V8Value value = v8ValueObject.get(key)) {
-                        map.put(keyString, toObject(value));
+                        map.put(keyString, toObject(value, depth + 1));
                     }
                 }
             });
@@ -114,8 +114,9 @@ public class JavetObjectConverter extends JavetPrimitiveConverter {
 
     @Override
     @CheckReturnValue
-    public <T extends V8Value> T toV8Value(V8Runtime v8Runtime, Object object) throws JavetException {
-        V8Value v8Value = super.toV8Value(v8Runtime, object);
+    protected <T extends V8Value> T toV8Value(
+            V8Runtime v8Runtime, Object object, final int depth) throws JavetException {
+        V8Value v8Value = super.toV8Value(v8Runtime, object, depth);
         if (v8Value != null && !(v8Value.isUndefined())) {
             return (T) v8Value;
         }
@@ -123,7 +124,7 @@ public class JavetObjectConverter extends JavetPrimitiveConverter {
             V8ValueMap v8ValueMap = v8Runtime.createV8ValueMap();
             Map<?, ?> mapObject = (Map<?, ?>) object;
             for (Object key : mapObject.keySet()) {
-                try (V8Value childV8Value = toV8Value(v8Runtime, mapObject.get(key))) {
+                try (V8Value childV8Value = toV8Value(v8Runtime, mapObject.get(key), depth + 1)) {
                     String childStringKey = key instanceof String ? (String) key : key.toString();
                     v8ValueMap.set(childStringKey, childV8Value);
                 }
@@ -133,7 +134,7 @@ public class JavetObjectConverter extends JavetPrimitiveConverter {
             V8ValueObject v8ValueObject = v8Runtime.createV8ValueObject();
             Map<?, ?> mapObject = (Map<?, ?>) object;
             for (Object key : mapObject.keySet()) {
-                try (V8Value childV8Value = toV8Value(v8Runtime, mapObject.get(key))) {
+                try (V8Value childV8Value = toV8Value(v8Runtime, mapObject.get(key), depth + 1)) {
                     String childStringKey = key instanceof String ? (String) key : key.toString();
                     v8ValueObject.set(childStringKey, childV8Value);
                 }
@@ -143,7 +144,7 @@ public class JavetObjectConverter extends JavetPrimitiveConverter {
             V8ValueSet v8ValueSet = v8Runtime.createV8ValueSet();
             Set<?> setObject = (Set<?>) object;
             for (Object item : setObject) {
-                try (V8Value childV8Value = toV8Value(v8Runtime, item)) {
+                try (V8Value childV8Value = toV8Value(v8Runtime, item, depth + 1)) {
                     v8ValueSet.add(childV8Value);
                 }
             }
@@ -151,7 +152,7 @@ public class JavetObjectConverter extends JavetPrimitiveConverter {
         } else if (object instanceof Collection) {
             V8ValueArray v8ValueArray = v8Runtime.createV8ValueArray();
             for (Object item : (Collection<?>) object) {
-                try (V8Value childV8Value = toV8Value(v8Runtime, item)) {
+                try (V8Value childV8Value = toV8Value(v8Runtime, item, depth + 1)) {
                     v8ValueArray.push(childV8Value);
                 }
             }
@@ -207,7 +208,7 @@ public class JavetObjectConverter extends JavetPrimitiveConverter {
         } else if (object.getClass().isArray()) {
             V8ValueArray v8ValueArray = v8Runtime.createV8ValueArray();
             for (Object item : (Object[]) object) {
-                try (V8Value childV8Value = toV8Value(v8Runtime, item)) {
+                try (V8Value childV8Value = toV8Value(v8Runtime, item, depth + 1)) {
                     v8ValueArray.push(childV8Value);
                 }
             }